apps/backend/requirements.txt
# ============================================
# Python Version: 3.12+
# ============================================

# ============================================
# FastAPI Framework & Server
# ============================================
fastapi==0.135.1
uvicorn[standard]==0.42.0

# ============================================
# Database ORM
# ============================================
sqlalchemy==2.0.48

# ============================================
# Data Validation & Settings
# ============================================
pydantic==2.12.5
pydantic-settings==2.13.1

# ============================================
# Configuration Management
# ============================================
python-dotenv==1.2.2

# ============================================
# Async File I/O
# ============================================
aiofiles==25.1.0

# ============================================
# Web Scraping & Browser Automation
# ============================================
playwright==1.58.0
scrapling==0.4.2
beautifulsoup4==4.14.3

# ============================================
# HTTP Clients
# ============================================
requests==2.32.5
httpx==0.28.1

# ============================================
# Task Scheduling
# ============================================
apscheduler==3.11.2

# ============================================
# Caching (Redis)
# ============================================
redis==7.3.0

# ============================================
# Testing & Development
# ============================================
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
fakeredis==2.34.1

# ============================================
# Code Quality & Linting
# ============================================
ruff==0.9.5
black==25.2.0
isort==5.13.2

# ============================================
# File Upload Support
# ============================================
python-multipart==0.0.22

# ============================================
# Meteorology & Scientific Computing
# ============================================
metpy==1.7.1
matplotlib==3.10.8
numpy==2.4.3
pillow==12.1.1

# ============================================
# AI/LLM Integration
# ============================================
anthropic==0.86.0
google-generativeai==0.8.6
